add missing fun of seeding tests for resource-based solution
