+++ core-it/src/test/kotlin/org/evomaster/core/problem/rest/service/resource/model/SimpleResourceSampler.kt
@@ -4,7 +4,6 @@ import io.swagger.parser.OpenAPIParser
 import org.evomaster.core.database.SqlInsertBuilder
 import org.evomaster.core.problem.rest.RestActionBuilderV3
 import org.evomaster.core.problem.rest.service.ResourceSampler
-import org.evomaster.core.search.Action
 import javax.annotation.PostConstruct
 
 class SimpleResourceSampler : ResourceSampler() {
