+++ core/src/main/kotlin/org/evomaster/core/problem/rest/service/AbstractRestSampler.kt
@@ -6,11 +6,12 @@ import org.evomaster.client.java.controller.api.dto.SutInfoDto
 import org.evomaster.core.EMConfig
 import org.evomaster.core.output.service.PartialOracles
 import org.evomaster.core.problem.httpws.service.HttpWsSampler
-import org.evomaster.core.problem.rest.OpenApiAccess
-import org.evomaster.core.problem.rest.RestActionBuilderV3
-import org.evomaster.core.problem.rest.RestIndividual
+import org.evomaster.core.problem.rest.*
+import org.evomaster.core.problem.rest.seeding.Parser
+import org.evomaster.core.problem.rest.seeding.postman.PostmanParser
 import org.evomaster.core.remote.SutProblemException
 import org.evomaster.core.remote.service.RemoteController
+import org.evomaster.core.search.tracer.Traceable
 import org.slf4j.Logger
 import org.slf4j.LoggerFactory
 import javax.annotation.PostConstruct
@@ -83,9 +84,28 @@ abstract class AbstractRestSampler : HttpWsSampler<RestIndividual>() {
         log.debug("Done initializing {}", AbstractRestSampler::class.simpleName)
     }
 
+    /**
+     * create AdHocInitialIndividuals
+     */
+    fun initAdHocInitialIndividuals(){
+        customizeAdHocInitialIndividuals()
 
-    abstract fun initAdHocInitialIndividuals()
+        // if test case seeding is enabled, add those test cases too
+        if (config.seedTestCases) {
+            val parser = getParser()
+            val seededTestCases = parser.parseTestCases(config.seedTestCasesPath)
+            adHocInitialIndividuals.addAll(seededTestCases.map { createIndividual(it) })
+        }
+    }
 
+    /**
+     * customize AdHocInitialIndividuals
+     */
+    abstract fun customizeAdHocInitialIndividuals()
+
+    /**
+     * post action after InitialIndividuals are crated
+     */
     open fun postInits(){
         //do nothing
     }
@@ -145,5 +165,16 @@ abstract class AbstractRestSampler : HttpWsSampler<RestIndividual>() {
 
     override fun hasSpecialInit(): Boolean {
         return !adHocInitialIndividuals.isEmpty() && config.probOfSmartSampling > 0
+    }
+
+    fun createIndividual(restCalls: MutableList<RestCallAction>): RestIndividual {
+        return RestIndividual(restCalls, SampleType.SMART, mutableListOf()//, usedObjects.copy()
+                ,trackOperator = if (config.trackingEnabled()) this else null, index = if (config.trackingEnabled()) time.evaluatedIndividuals else Traceable.DEFAULT_INDEX)
+    }
+
+    private fun getParser(): Parser {
+        return when(config.seedTestCasesFormat) {
+            EMConfig.SeedTestCasesFormat.POSTMAN -> PostmanParser(seeAvailableActions().filterIsInstance<RestCallAction>(), swagger)
+        }
     }
 }
